requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-bootstrap5==23.3
django-elasticsearch-dsl==8.0
django-phonenumber-field==7.2.0
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
idna==3.6
phonenumbers==8.13.26
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
